Ok clone it, but at least send some trafic my way
I took this down cause it was being cloned like mad, I know its very useful as it gives a fast way to search API and the source in the .lua file right down to the line number of the function. It would be nice if you gave some credit to me but hey, when in rome right.